Network.ABCI.Server

Documentation

defaultLocalSettings :: ServerSettings

Default ABCI app network settings for serving on localhost at the standard port.

serveAppWith :: ServerSettings -> (AppData -> IO ()) -> App IO -> IO ()

Serve an ABCI application with custom ServerSettings and a custom action to perform on acquiring the socket resource.

serveApp :: App IO -> IO ()

Serve an ABCI application with default local ServerSettings and a no-op on acquiring the socket resource.