Network.ABCI.Server.App

Documentation

newtype App m

Application type that represents a well typed application, i.e. a function from a typed Request to a typed Response.

Constructors

App

Fields

unApp :: forall (t :: MessageType). Request t -> m (Response t)

runApp :: forall m. Applicative m => App m -> LPByteStrings -> m LPByteStrings

Compiles App down to AppBS

transformApp :: (forall (t :: MessageType). m (Response t) -> g (Response t)) -> App m -> App g

Transform an application from running in a custom monad to running in IO.

withProto :: (forall (t :: MessageType). Request t -> a) -> Request'Value -> a

type Middleware m = App m -> App m

Middleware is a component that sits between the server and application. It can do such tasks as logging or response caching. What follows is the general definition of middleware, though a middleware author should feel free to modify this.

data MessageType

Used to parametrize Request and Respone types

Constructors

MTEcho

MTFlush

MTInfo

MTSetOption

MTInitChain

MTQuery

MTBeginBlock

MTCheckTx

MTDeliverTx

MTEndBlock

MTCommit

demoteRequestType :: forall (t :: MessageType). Request t -> MessageType

msgTypeKey :: MessageType -> String

data Request (m :: MessageType) :: * where

Constructors

RequestEcho :: Echo -> Request MTEcho

RequestInfo :: Info -> Request MTInfo

RequestSetOption :: SetOption -> Request MTSetOption

RequestQuery :: Query -> Request MTQuery

RequestCheckTx :: CheckTx -> Request MTCheckTx

RequestFlush :: Flush -> Request MTFlush

RequestInitChain :: InitChain -> Request MTInitChain

RequestBeginBlock :: BeginBlock -> Request MTBeginBlock

RequestDeliverTx :: DeliverTx -> Request MTDeliverTx

RequestEndBlock :: EndBlock -> Request MTEndBlock

RequestCommit :: Commit -> Request MTCommit

hashRequest :: forall (t :: MessageType). Request t -> HexString

data Response (m :: MessageType) :: * where

Constructors

ResponseEcho :: Echo -> Response MTEcho

ResponseFlush :: Flush -> Response MTFlush

ResponseInfo :: Info -> Response MTInfo

ResponseSetOption :: SetOption -> Response MTSetOption

ResponseInitChain :: InitChain -> Response MTInitChain

ResponseQuery :: Query -> Response MTQuery

ResponseBeginBlock :: BeginBlock -> Response MTBeginBlock

ResponseCheckTx :: CheckTx -> Response MTCheckTx

ResponseDeliverTx :: DeliverTx -> Response MTDeliverTx

ResponseEndBlock :: EndBlock -> Response MTEndBlock

ResponseCommit :: Commit -> Response MTCommit

ResponseException :: forall (m :: MessageType). Exception -> Response m

newtype LPByteStrings

ByteString which contains multiple length prefixed ByteStrings

Constructors

LPByteStrings

Fields

unLPByteStrings :: ByteString

decodeLengthPrefix :: LPByteStrings -> Either DecodeError [ByteString]

Decodes varlength-prefixed ByteString into ByteStrings

encodeLengthPrefix :: [ByteString] -> LPByteStrings

Encodes ByteStrings into varlength-prefixed ByteString